Extract capture key reset hook to satisfy size gate
